change Bold Attributes logic
I realized that saving csv file to data folder would not work because that data folder is ephemeral. It will be deleted after the app closes, so I needed to create the default CSV if it doesn't exist, similar to creating the config.ini file. Then if the CSV file is found upon opening the app, read the CSV file and save back to it when needed.