Changed design of WIS Creds UI
Because I changed the auth logic from automating the username and password and login everything to instead having the user login once manually and then store the cookie in a file to be reused for future sessions.

Had to rework how the coachid and cookie data was stored and retrieved.

Added a login button and a checkmark/x-icon indicating whether a cookie is stored for user or not.