trying to resolve db is locked errors when update considering
For some reason when doing update considering action, when trying to bind the query and execute, an error is returned about the DB being locked.

Right now the only way I have figured out how to work around this is to add db.transaction() and db.commit() to the update considering section.

I have not been able to confirm that new data is actually getting written because I don't beleive the advanced search page is getting updated properly.