Remove scholarship counts from considering for signed recruits
Once a recruit was signed, the considering field would never be updated again, and it was storing the number of scholarships open at the time the update was run. So one signed recruit might show the school had 5 open scholarships and another recruit might show the school has 2 open scholarships.

In order to avoid this situation and prevent confusion, when a player is reported as signed, the considering field will only store the school name, the coachid, and the division.

Unsigned recruits will show all the additional info in considering field including miles away, total scholarships, and remaining open scholarships.